% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{cs_academic_years}
\alias{cs_academic_years}
\alias{cs_school_years}
\alias{cs_response_types}
\alias{cs_response_value_groups}
\alias{cs_person_types}
\alias{cs_cohorts}
\alias{cs_school_types}
\title{Get CS meta information}
\usage{
cs_academic_years(token = NULL)

cs_school_years(token = NULL)

cs_response_types(token = NULL)

cs_response_value_groups(token = NULL)

cs_person_types(token = NULL)

cs_cohorts(token = NULL)

cs_school_types(token = NULL)
}
\arguments{
\item{token}{API token}
}
\value{
academic years database
}
\description{
Get CS meta information
}
