% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{cs_endpoint}
\alias{cs_endpoint}
\title{Pull from meta endpoints}
\usage{
cs_endpoint(token = NULL, endpoint = "")
}
\arguments{
\item{token}{token}

\item{endpoint}{endpoint}
}
\description{
Pull from meta endpoints
}
