% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{cs_schools}
\alias{cs_schools}
\title{Pull from CS API - schools/school endpoint}
\usage{
cs_schools(token = NULL, id = NA, include = NA)
}
\arguments{
\item{token}{token}

\item{id}{school ID in CS database. If NA, all schools are pulled.}

\item{include}{Related objects to include. Valid values are: 'class', 'student', 'teacher'. Comma separated list of objects to include.}
}
\value{
data.frame of pulled information
}
\description{
Pull from CS API - schools/school endpoint
}
