% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_df}
\alias{flatten_df}
\title{flatten a data frame with (multiple) list elements}
\usage{
flatten_df(df)
}
\arguments{
\item{df}{data frame}
}
\value{
flat data frame (long format)
}
\description{
flatten a data frame with (multiple) list elements
}
