% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{survey_response_to_wide}
\alias{survey_response_to_wide}
\title{Convert 'flattened' survey to wide format.}
\usage{
survey_response_to_wide(survey)
}
\arguments{
\item{survey}{flattened survey.}
}
\value{
survey in the wide format.
}
\description{
Convert 'flattened' survey to wide format.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
survey = cs_surveys_responses(id = 'ENTER_ID_HERE', output_type = 'original' )
survey_wide = survey_response_to_wide(survey)
\dontshow{\}) # examplesIf}
}
