% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{cs_schools}
\alias{cs_schools}
\title{Query school endpoints}
\usage{
cs_schools(token = NULL, id = NA, include = NA)
}
\arguments{
\item{token}{API token}

\item{id}{school ID in CS database. If NA, all schools are pulled.}

\item{include}{Character vector or objects to include. Accepted values are: 'class', 'student', 'teacher'.}
}
\value{
data.frame of pulled information
}
\description{
A functions to query school-related information from the Cohort Studies API.
}
