% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{cs_surveys}
\alias{cs_surveys}
\alias{cs_surveys_responses}
\title{Query survey endpoints}
\usage{
cs_surveys(token = NULL, id = NA, cohortId = NA)

cs_surveys_responses(
  token = NULL,
  id = NA,
  respondentIds = "",
  teacherIds = "",
  schoolIds = "",
  classIds = "",
  output_type = "wide",
  verbose = F
)
}
\arguments{
\item{token}{token}

\item{id}{survey ID in CS database. If NA, all surveys are pulled.}

\item{cohortId}{verbose}

\item{respondentIds}{verbose}

\item{teacherIds}{verbose}

\item{schoolIds}{verbose}

\item{classIds}{verbose}

\item{output_type}{output_type}

\item{verbose}{verbose}
}
\value{
data.frame of pulled information
}
\description{
Query survey endpoints
}
