% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{cs_surveys}
\alias{cs_surveys}
\alias{cs_surveys_responses}
\title{Pull from CS API - surveys / surveys\{id\} / survey\{id\}/response endpoints}
\usage{
cs_surveys(token = NULL, id = NA)

cs_surveys_responses(token = NULL, id = NA)
}
\arguments{
\item{token}{token}

\item{id}{survey ID in CS database. If NA, all schools are pulled.}
}
\value{
data.frame of pulled information
}
\description{
Pull from CS API - surveys / surveys\{id\} / survey\{id\}/response endpoints
}
