% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_api.R
\name{get_query}
\alias{get_query}
\title{Pull from OME database}
\usage{
get_query(url_path, filter = "", KEY = "")
}
\arguments{
\item{url_path}{the API endpoint}

\item{filter}{Addition objects to pass via include. (not currently in use)}

\item{KEY}{API token}
}
\value{
data frame of pulled information
}
\description{
Pull from OME database
}
